Conflicts:
	xkill-physics/BulletPhysicsComponent.cpp
	xkill-physics/BulletPhysicsComponent.h
	xkill-physics/physicsObject.cpp
	xkill-physics/physicsObject.h

merged with dev to get latest changes to attributes

+++ xkill-input/InputAxisObject.cpp
@@ -33,7 +33,7 @@ void InputAxisObject::SetValue(float value)
 
 void InputAxisObject::SetValue(int value)
 {
-	prevValue_ = value;
+	prevValue_ = static_cast<float>(value);
 	value_ = formatValue(value);
 }
